Remove VentanaBenchmark from SpecialStain use from Surgical, refactor BlockCount to Dashboard
