--- YellowstonePathology/Business/SpecialStain/VentanaBenchMarkCollection.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Collections.ObjectModel;
-
-namespace YellowstonePathology.Business.SpecialStain
-{
-    public class VentanaBenchMarkCollection : ObservableCollection<VentanaBenchMark>
-    {
-        public VentanaBenchMarkCollection()
-        { }
-    }
-}
